refactor UseMemo to own file
